# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise Ticketing, Project Management & Workflow System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: department
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── DEPARTMENTS ──────────────────────────────────────────────────────────────
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── ROLES ────────────────────────────────────────────────────────────────────
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    patch:
      operationId: updateRole
      tags: [roles]
      summary: Update role
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete role
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── CATEGORIES ───────────────────────────────────────────────────────────────
...
  /todos:
    get:
      operationId: listTodos
      tags: [todos]
      summary: List todos
      parameters:
        - name: type
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: assignedTo
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      operationId: createTodo
      tags: [todos]
      summary: Create todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    patch:
      operationId: updateTodo
      tags: [todos]
      summary: Update todo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    delete:
      operationId: deleteTodo
      tags: [todos]
      summary: Delete todo
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /todos/{id}/complete:
    patch:
      operationId: completeTodo
      tags: [todos]
      summary: Mark todo complete/incomplete
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoComplete"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}/convert-ticket:
    post:
      operationId: convertTodoToTicket
      tags: [todos]
      summary: Convert todo to ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  # ── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, employeeCode, name, email, role, status, createdAt]
      properties:
        id: { type: integer }
        employeeCode: { type: string }
        name: { type: string }
        email: { type: string }
        mobile: { type: ["string", "null"] }
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        designation: { type: ["string", "null"] }
        role: { type: string }
        roleId: { type: ["integer", "null"] }
        reportingManagerId: { type: ["integer", "null"] }
        reportingManagerName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        status: { type: string }
        lastLogin: { type: ["string", "null"] }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [employeeCode, name, email, role]
      properties:
        employeeCode: { type: string }
        name: { type: string }
        email: { type: string }
        mobile: { type: string }
        departmentId: { type: integer }
        designation: { type: string }
        role: { type: string }
        roleId: { type: integer }
        reportingManagerId: { type: integer }
        password: { type: string }
        status: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        mobile: { type: string }
        departmentId: { type: integer }
        designation: { type: string }
        role: { type: string }
        roleId: { type: integer }
        reportingManagerId: { type: integer }
        status: { type: string }

    # ── DEPARTMENT SCHEMAS ──────────────────────────────────────────────────
    Department:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        headId: { type: ["integer", "null"] }
        headName: { type: ["string", "null"] }
        userCount: { type: integer }
        createdAt: { type: string }
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        headId: { type: integer }

    # ── ROLE SCHEMAS ────────────────────────────────────────────────────────
    Role:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        level: { type: integer }
        createdAt: { type: string }
    RoleInput:
      type: object
      required: [name, level]
      properties:
        name: { type: string }
        description: { type: string }
        level: { type: integer }

    # ── CATEGORY SCHEMAS ────────────────────────────────────────────────────
...
    Ticket:
      type: object
      required: [id, ticketNo, subject, status, priority, createdAt]
      properties:
        id: { type: integer }
        ticketNo: { type: string }
        subject: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        category: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        type: { type: string }
        createdById: { type: integer }
        createdByName: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
        assignedToName: { type: ["string", "null"] }
        assignedToAvatar: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        resolvedAt: { type: ["string", "null"] }
        closedAt: { type: ["string", "null"] }
        pendingDays: { type: ["integer", "null"] }
        slaBreached: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Todo:
      type: object
      required: [id, title, status, priority, type, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        type: { type: string }
        dueDate: { type: ["string", "null"] }
        reminderAt: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
        assignedToName: { type: ["string", "null"] }
        createdById: { type: integer }
        completedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    TodoInput:
      type: object
      required: [title, priority, type]
      properties:
        title: { type: string }
        description: { type: string }
        priority: { type: string }
        type: { type: string }
        dueDate: { type: string }
        reminderAt: { type: string }
        assignedToId: { type: integer }
    TodoUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        priority: { type: string }
        dueDate: { type: string }
        reminderAt: { type: string }
    TodoComplete:
      type: object
      required: [completed]
      properties:
        completed: { type: boolean }

    # ── NOTIFICATION SCHEMAS ────────────────────────────────────────────────
...