# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Ticketing, Project Management & Workflow System API
 * OpenAPI spec version: 0.1.0
 */

export interface Role {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  level?: number;
  createdAt: string;
}