# lib/api-zod/src/generated/types/todoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Ticketing, Project Management & Workflow System API
 * OpenAPI spec version: 0.1.0
 */

export interface TodoInput {
  title: string;
  description?: string;
  priority: string;
  type: string;
  dueDate?: string;
  reminderAt?: string;
  assignedToId?: number;
}