# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Ticketing, Project Management & Workflow System API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  employeeCode: string;
  name: string;
  email: string;
  /** @nullable */
  mobile?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  designation?: string | null;
  role: string;
  /** @nullable */
  roleId?: number | null;
  /** @nullable */
  reportingManagerId?: number | null;
  /** @nullable */
  reportingManagerName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  status: string;
  /** @nullable */
  lastLogin?: string | null;
  createdAt: string;
}