# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Enterprise Ticketing, Project Management & Workflow System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  mobile?: string;
  departmentId?: number;
  designation?: string;
  role?: string;
  roleId?: number;
  reportingManagerId?: number;
  status?: string;
}